Initial commit of Dokucuoglu Yazilim project files
This commit includes the initial version of the Dokucuoglu Yazilim project, which aims to facilitate various tasks for the Dokucuoglu company. The project currently consists of a main application built with Tkinter for managing and comparing invoices using Excel files. The application features a user-friendly interface and customizable styles. Future developments will focus on enhancing functionality and improving user experience.